Moved main.go in src so that fyne-cross could compile. Added metadata file
