+++ .gitignore
@@ -1,1 +1,3 @@
 credentials.env
+main.exe
+fyne-cross
