+++ metadata.json
@@ -1,0 +1,5 @@
+{
+    "name": "TT Companion",
+    "appID": "com.TTCompanion",
+    "version": "1.2.0"
+  }
